refactor: inject data
