+++ src/api/database/seeds/createPermissions.ts
@@ -1,7 +1,10 @@
-import { roleActions, apiKeyActions } from '../../persistence/permission/permission.entity.js';
 import { BypassPrismaType } from '../prisma/client.js';
 
-export const createPermissions = async (prisma: BypassPrismaType): Promise<void> => {
+export const createPermissions = async (
+  prisma: BypassPrismaType,
+  roleActions: { [key: string]: string[] },
+  apiKeyActions: { [key: string]: string[] }
+): Promise<void> => {
   await prisma.permission.createMany({
     data: [
       // role permissions
