+++ src/api/database/seeds/createProjects.ts
@@ -1,26 +1,11 @@
-import { v4 } from 'uuid';
+import { Project } from '@prisma/client';
 import { BypassPrismaType } from '../prisma/client.js';
 
-export const enProject = v4();
-export const jpProject = v4();
-
-export const createProjects = async (prisma: BypassPrismaType): Promise<void> => {
+export const createProjects = async (
+  prisma: BypassPrismaType,
+  projects: Omit<Project, 'createdAt' | 'updatedAt'>[]
+): Promise<void> => {
   await prisma.project.createMany({
-    data: [
-      {
-        id: enProject,
-        subdomain: 'en',
-        sourceLanguage: 'en-us',
-        iconUrl: null,
-        name: 'EN Project',
-      },
-      {
-        id: jpProject,
-        subdomain: 'ja',
-        sourceLanguage: 'ja',
-        iconUrl: null,
-        name: 'JA Project',
-      },
-    ],
+    data: projects,
   });
 };
