+++ src/api/database/seeds/createUsers.ts
@@ -3,18 +3,14 @@ import { v4 } from 'uuid';
 import { oneWayHash } from '../../utilities/oneWayHash.js';
 import { BypassPrismaType } from '../prisma/client.js';
 
-// Fixed to avoid mismatch between session and user id.
-export const adminUser = '1bc49902-bee5-4c90-b87a-3254d5c7e504';
-export const editorUser = '513a1cb6-68d9-4618-bc6d-4bb2c3b54048';
-export const contributorUser = 'ea23ce3d-b236-41b8-9696-7b2a6377451f';
-export const viewerUser = 'f3d75e36-387b-459c-a227-83c7a0c54a1a';
-
 export const createUsers = async (
   prisma: BypassPrismaType,
   users: {
     id: string;
     email: string;
     password: string;
+    confirmationToken: string | null;
+    confirmedAt: Date | null;
     userProjects: { projectId: string; roleId: string }[];
   }[]
 ): Promise<void> => {
@@ -27,8 +23,10 @@ export const createUsers = async (
         avatarUrl: faker.image.avatarGitHub(),
         password: await oneWayHash(user.password),
         isActive: true,
-        confirmationToken: v4(),
-        confirmedAt: new Date(),
+        confirmationToken: user.confirmationToken,
+        confirmedAt: user.confirmedAt,
+        // confirmationToken: v4(),
+        // confirmedAt: new Date(),
         userProjects: {
           create: user.userProjects.map((userProject) => ({
             id: v4(),
